se agrega la validacion en marca y categoria para que si esta en uso no se puedan elimiar.

+++ tp-winform-equipo-14B/frmEliminarCategoria.cs
@@ -43,6 +43,13 @@ namespace tp_winform_equipo_14B
                 try
                 {
                     CategoriaNegocio catNegocio = new CategoriaNegocio();
+                    //
+                    if (catNegocio.ExisteCategoria(idCategoria))
+                    {
+                        MessageBox.Show("La categoría esta en uso. No se puede borrar.");
+                        return;
+                    }
+                    //
                     catNegocio.eliminarCategoria(idCategoria);
 
                     ListCatElim = catNegocio.listar();
